cleaning up, styles, hooks
