+++ src/App.js
@@ -1,41 +1,37 @@
 /* eslint-disable no-undef */
-import React, { useState, useEffect } from "react";
+import React from "react";
 import catForkImg from "./images/catForkImg_square.png";
-import { CatPic } from "./components/CatPicComponent.js";
+import { CatPicComponent } from "./Features/CatPic/CatPicComponent";
 import "./App.css";
 import RecipeSearchBox from "./Features/Recipes/Search/RecipeSearchBox";
 import RecipeList from "./Features/Recipes/components/RecipeList";
 import firebase from "firebase";
 import firebaseConfig from "./Features/Firebase/firebaseConfig";
+// TODO: remove bootstrap
 import "bootstrap/dist/css/bootstrap.min.css";
 firebase.initializeApp(firebaseConfig);
 
 export const App = () => {
-  const [catPicUrl, setCatPicUrl] = useState("");
-
-  // getCatPic(setCatPicUrl);
-  useEffect(() => {
-    if (!catPicUrl) {
-      console.log("setting CatPic in UseEffect");
-      getCatPic(setCatPicUrl);
-    }
-  });
+  
 
   return (
+    // TODO: change text for mobile
+    // TODO: Cats eating salami background marquee, have the images
     <div className="app">
-      <h3 className="app-intro">
-        {" "}
-        This is a recipe search app based on ingredients you enter. Please
-        separate them by a comma and enter at least two (2).{" "}
+      
+      <h3 className="appIntroText">
+        This is a recipe search app based on ingredients you enter. <br />
+        <a href="https://thecatapi.com"> 
+          And Powered by Cats
+          </a>
       </h3>
-      <div className="catPicContainer">
+      <div className="main">
         <div>
           <img className="catforkLogo" src={catForkImg} alt="logo" />
         </div>
-        <CatPic catUrl={catPicUrl} />
-        <button className="btn-class" onClick={() => getCatPic(setCatPicUrl)}>
-          New Cat
-        </button>
+        
+        <CatPicComponent />
+        
       </div>
 
       <RecipeSearchBox />
@@ -46,14 +42,5 @@ export const App = () => {
 };
 
 export default App;
-async function getCatPic(setCatPicUrl) {
-  var response = await fetch(
-    "https://thecatapi.com/api/images/get?format=src&size=med"
-  );
-  // eslint-disable-next-line no-console
-  console.log("hitting getnew cat", response);
-  setCatPicUrl(response.url);
-  // this.setState({
-  //   catUrl: response.url
-  // });
-}
+
+
