+++ src/Features/CatPic/CatPic.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const CatPic = props => {
+  return (
+    <div className="catPic">
+      <img className="catFull" src={props.catUrl} alt="logo" />
+    </div>
+  );
+};
